Created plantcv.learn subpackage
Moved naive Bayes functions to plantcv.learn subpackage
