+++ plantcv/learn/__init__.py
@@ -1,0 +1,3 @@
+from naive_bayes import naive_bayes
+
+__all__ = ["naive_bayes"]
